refactor: 멤버 서비스 인터페이스 삭제 및 리팩토링 #16
