+++ src/main/java/com/sm/project/web/controller/member/MemberController.java
@@ -12,7 +12,6 @@ import com.sm.project.apiPayload.code.ErrorReasonDTO;
 import com.sm.project.apiPayload.code.status.SuccessStatus;
 import com.sm.project.converter.member.MemberConverter;
 import com.sm.project.domain.member.Member;
-import com.sm.project.service.member.MemberCommandService;
 import com.sm.project.service.member.MemberQueryService;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -37,7 +36,6 @@ import java.io.UnsupportedEncodingException;
 public class MemberController {
 
     private final MemberService memberService;
-    private final MemberCommandService memberCommandService;
     private final MemberQueryService memberQueryService;
     private final MailService mailService;
 
@@ -72,7 +70,7 @@ public class MemberController {
                     content = @Content(schema = @Schema(implementation = ErrorReasonDTO.class))),
     })
     public ResponseDTO<MemberResponseDTO.JoinResultDTO> join(@RequestBody @Valid MemberRequestDTO.JoinDTO request) {
-        Member newMember = memberCommandService.joinMember(request);
+        Member newMember = memberService.joinMember(request);
         return ResponseDTO.of(SuccessStatus._OK, MemberConverter.toJoinResultDTO(newMember));
     }
 
@@ -84,7 +82,7 @@ public class MemberController {
                     content = @Content(schema = @Schema(implementation = ErrorReasonDTO.class)))
     })
     public ResponseDTO checkNickname(@RequestBody @Valid MemberRequestDTO.NicknameDTO request) {
-        if (memberCommandService.isDuplicate(request)) {
+        if (memberService.isDuplicate(request)) {
             throw new MemberHandler(ErrorStatus.MEMBER_NICKNAME_DUPLICATE);
         }
         return ResponseDTO.onSuccess("닉네임 중복이 아닙니다.");
@@ -100,7 +98,7 @@ public class MemberController {
                     content = @Content(schema = @Schema(implementation = ErrorReasonDTO.class))),
     })
     public ResponseDTO<MemberResponseDTO.EmailResultDTO> findEmail(@RequestBody @Valid MemberRequestDTO.FindEmailDTO request) {
-        memberCommandService.verifySms(request.getPhone(), request.getCertificationCode());
+        memberService.verifySms(request.getPhone(), request.getCertificationCode());
         Member member = memberQueryService.findEmail(request.getPhone());
         return ResponseDTO.of(SuccessStatus._OK, MemberConverter.toEmailResultDTO(member));
     }
@@ -108,7 +106,7 @@ public class MemberController {
     @PostMapping("/send")
     @Operation(summary = "본인인증 문자 전송 API", description = "본인인증을 위한 인증번호 문자를 보내는 API입니다.")
     public ResponseDTO sendSMS(@RequestBody MemberRequestDTO.SmsDTO request) {
-        memberCommandService.sendSms(request);
+        memberService.sendSms(request);
         return ResponseDTO.onSuccess("인증문자 전송 성공");
     }
 
@@ -119,7 +117,7 @@ public class MemberController {
                     content = @Content(schema = @Schema(implementation = ErrorReasonDTO.class))),
     })
     public ResponseDTO sendEmail(@RequestBody @Valid MemberRequestDTO.FindPasswordDTO request) throws MessagingException, UnsupportedEncodingException {
-        memberCommandService.sendEmail(request);
+        memberService.sendEmail(request);
         return ResponseDTO.onSuccess("메일 전송 성공");
     }
 
@@ -132,7 +130,7 @@ public class MemberController {
                     content = @Content(schema = @Schema(implementation = ErrorReasonDTO.class)))
     })
     public ResponseDTO<?> resetPassword(@RequestParam String resetToken, @RequestBody @Valid MemberRequestDTO.PasswordDTO request) {
-        memberCommandService.resetPassword(resetToken, request);
+        memberService.resetPassword(resetToken, request);
         return ResponseDTO.onSuccess("비밀번호 재설정 성공");
     }
 }
